\name{InteractiveGraphs-package}
\alias{InteractiveGraphs-package}
\alias{InteractiveGraphs}
\docType{package}
\title{
\packageTitle{InteractiveGraphs}
}
\description{
\packageDescription{InteractiveGraphs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{InteractiveGraphs}
\packageIndices{InteractiveGraphs}

}
\author{
\packageAuthor{InteractiveGraphs}

Maintainer: \packageMaintainer{InteractiveGraphs}
}
\references{

}
\keyword{ package }
\seealso{

}
\examples{

}
